Refactored code to isolate PCA specific stuff to cloobvis.py
